+++ src/pages/programs/form/eventProgram/EventProgramFormContents.tsx
@@ -9,9 +9,11 @@ import {
     useSectionedFormContext,
     useSyncSelectedSectionWithScroll,
 } from '../../../../lib'
+import { AccessAndSharingFormContents } from '../common/AccessAndSharingFormContents'
 import { SetupFormContents } from '../common/SetupFormContents'
 import { StageDataFormContents } from '../programStage/StageDataFormContents'
 import { StageFormFormContents } from '../programStage/StageFormFormContents'
+import { EventProgramCustomizationFormContents } from './EventProgramCustomizationFormContents'
 import { EventProgramFormDescriptor } from './eventProgramFormDescriptor'
 
 export const EventProgramFormContents = () => {
@@ -24,9 +26,9 @@ export const EventProgramFormContents = () => {
                 name={descriptor.getSection('programDetails').name}
                 isTrackerProgram={false}
             />
-            {/*<ProgramCustomizationFormContents*/}
-            {/*    name={descriptor.getSection('programCustomization').name}*/}
-            {/*/>*/}
+            <EventProgramCustomizationFormContents
+                name={descriptor.getSection('programCustomization').name}
+            />
             {/*<EnrollmentSettingsFormContents*/}
             {/*    name={descriptor.getSection('enrollmentSettings').name}*/}
             {/*/>*/}
@@ -47,9 +49,10 @@ export const EventProgramFormContents = () => {
             {/*<ProgramNotificationsFormContents*/}
             {/*    name={descriptor.getSection('programNotifications').name}*/}
             {/*/>*/}
-            {/*<AccessAndSharingFormContents*/}
-            {/*    name={descriptor.getSection('accessAndSharing').name}*/}
-            {/*/>*/}
+            <AccessAndSharingFormContents
+                name={descriptor.getSection('accessAndSharing').name}
+                showStageAccess={false}
+            />
             <CustomAttributesSection
                 schemaSection={SCHEMA_SECTIONS.program}
                 sectionedLayout